Support removal of undesired headers from the original email
Depending on the end user configuration, some headers such as the
"Attachments:" header may appear in the rendered original email and
then end up in the separator preceding the original email.  To offer a
solution without having to alter the end-user configuration, this
patch introduces the `org-msg-undesirable-headers' customization
variable. This variable list the undesirable headers name which will
be deleted by the `org-msg-improve-reply-header' function.

This patch addresses #119.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>